+++ app/src/main/ui/src/components/users/UserForm.tsx
@@ -1,12 +1,15 @@
-import {deprecated_Form as Form} from "@cloudogu/ces-theme-tailwind";
+import {deprecated_Form as Form, Details} from "@cloudogu/ces-theme-tailwind";
 import {Button, H2, ListWithSearchbar} from "@cloudogu/deprecated-ces-theme-tailwind";
 import {TrashIcon} from "@heroicons/react/24/outline";
+import React from "react";
+import {twMerge} from "tailwind-merge";
 import {t} from "../../helpers/i18nHelpers";
 import {useConfirmation} from "../../hooks/useConfirmation";
 import {Prompt} from "../../hooks/usePrompt";
 import useUserFormHandler from "../../hooks/useUserFormHandler";
-import { GroupsService} from "../../services/Groups";
+import {GroupsService} from "../../services/Groups";
 import {ConfirmationDialog} from "../ConfirmationDialog";
+import HelpLink from "../helpLink";
 import type {Group} from "../../services/Groups";
 import type {User} from "../../services/Users";
 import type {NotifyFunction, UseFormHandlerFunctions} from "@cloudogu/deprecated-ces-theme-tailwind";
@@ -164,5 +167,16 @@ export default function UserForm<T extends User>(props: UserFormProps<T>) {
             </>
             : <></>
         }
+        <hr/>
+        <Details className={twMerge("py-2")}>
+            <Details.Summary>
+                <Details.Summary.Arrow/>
+                {t("users.steps.title")}
+            </Details.Summary>
+            <Details.Content>
+                {t("users.steps.text")}
+                <HelpLink/>
+            </Details.Content>
+        </Details>
     </>;
 }